#117 Rename battle/engagement.rs to battle/state.rs
