+++ src/game/engagement/battle/state.rs
